src: make a raytracer to test and debug the bounding boxes
index: incomplete, only bounding box nodes exists, index must be
completed with triangle nodes.

bounding box: must aggregate non-volume nodes into separate array
(currently the only non-bounding box node must be triangle to make the
prototype easier to implement in WebGPU, later we might add support for
other shapes).

python/numpy: depth buffer must be matched atomically. This allows to
update a pixel only if the depth value is smaller than the stored value
in the depth buffer for that pixel. Unfortunately, such atomic operation
is not possible in numpy (and it is not actually possible in WebGPU, but
if the depth buffer float are positive only, the integer interpretation
of their bits will have the same order). To mitigate this, python
implementations does two separate matching: using pandas to group
current dispatch of jobs by pixel and select minimum depth, then
separately compare the minimum depth of each pixel in the current
dispatch to value in the depth buffer (which is set by previous
dispatch). This is more complicated but the fastest way to do it in
python (avoiding per-pixel loops). When rewritten to WebGPU depth
testing should be simpler.