raytracing: done
Will move the index generation to make-index.py and then store that as
a file. Then a numpy can read that file directly (instead of parsing
obj text file again) and do the raytracing without repeating the index.

TODO: Must be able to produce index from existing index and
transformation and potentially additional data to be indexed.