src: refactoring of index and separation of actions
Separation of actions: indexing and raytracing are now separate actions.
Output file: indexing generate an output file, so the result of the
indexing can be stored instead of recomputed every time.
Logger: A separate classes to provide output after various operations of
the indexer (and later - the raytracer).
Index: We notices that the node data type is known in advance, so now
the node array can be allocated at the store. Every node contains at
least one node, thus the number of nodes cannot exceed the number of
triangles. This allows us to allocate all necessary memory upfront and
only modify the existing memory.
Index: The pandas.groupby automatically does sort=True by default, which
is Nlog(N). With sort=False, the grouping performance should be better,
especially in cases where we pre-sort the data.